Externalize actions, keys, persistence, and people
