+++ src/actions.js
@@ -1,0 +1,6 @@
+/** @const */   export var SET_ACTION                = '$set';
+/** @const */   export var SET_ONCE_ACTION           = '$set_once';
+/** @const */   export var UNSET_ACTION              = '$unset';
+/** @const */   export var ADD_ACTION                = '$add';
+/** @const */   export var APPEND_ACTION             = '$append';
+/** @const */   export var UNION_ACTION              = '$union';
