+++ src/persistence-keys.js
@@ -1,0 +1,11 @@
+/** @const */   export var SET_QUEUE_KEY             = '__mps';
+/** @const */   export var SET_ONCE_QUEUE_KEY        = '__mpso';
+/** @const */   export var UNSET_QUEUE_KEY           = '__mpus';
+/** @const */   export var ADD_QUEUE_KEY             = '__mpa';
+/** @const */   export var APPEND_QUEUE_KEY          = '__mpap';
+/** @const */   export var UNION_QUEUE_KEY           = '__mpu';
+/** @const */   export var EVENT_TIMERS_KEY          = '__timers';
+// This key is deprecated, but we want to check for it to see whether aliasing is allowed.
+/** @const */   export var PEOPLE_DISTINCT_ID_KEY    = '$people_distinct_id';
+/** @const */   export var ALIAS_ID_KEY              = '__alias';
+/** @const */   export var CAMPAIGN_IDS_KEY          = '__cmpns';
